+++ src/Factories/Factory.php
@@ -18,37 +18,34 @@
 
 namespace CloudCreativity\JsonApi\Factories;
 
+use CloudCreativity\JsonApi\Contracts\ContainerInterface;
 use CloudCreativity\JsonApi\Contracts\Encoder\SerializerInterface;
 use CloudCreativity\JsonApi\Contracts\Factories\FactoryInterface;
-use CloudCreativity\JsonApi\Contracts\Http\Requests\RequestInterpreterInterface;
+use CloudCreativity\JsonApi\Contracts\Object\DocumentInterface;
 use CloudCreativity\JsonApi\Contracts\Repositories\ErrorRepositoryInterface;
-use CloudCreativity\JsonApi\Contracts\Store\ContainerInterface as AdapterContainerInterface;
 use CloudCreativity\JsonApi\Contracts\Store\StoreInterface;
 use CloudCreativity\JsonApi\Contracts\Validators\QueryValidatorInterface;
 use CloudCreativity\JsonApi\Encoder\Encoder;
 use CloudCreativity\JsonApi\Http\Client\GuzzleClient;
 use CloudCreativity\JsonApi\Http\Query\ValidationQueryChecker;
-use CloudCreativity\JsonApi\Http\Requests\RequestFactory;
+use CloudCreativity\JsonApi\Http\Requests\InboundRequest;
 use CloudCreativity\JsonApi\Http\Responses\ErrorResponse;
 use CloudCreativity\JsonApi\Http\Responses\Response;
 use CloudCreativity\JsonApi\Object\Document;
 use CloudCreativity\JsonApi\Pagination\Page;
 use CloudCreativity\JsonApi\Repositories\CodecMatcherRepository;
 use CloudCreativity\JsonApi\Repositories\ErrorRepository;
-use CloudCreativity\JsonApi\Store\Container as AdapterContainer;
 use CloudCreativity\JsonApi\Store\Store;
 use CloudCreativity\JsonApi\Utils\Replacer;
 use CloudCreativity\JsonApi\Validators\ValidatorErrorFactory;
 use CloudCreativity\JsonApi\Validators\ValidatorFactory;
-use Neomerx\JsonApi\Contracts\Document\ErrorInterface;
 use Neomerx\JsonApi\Contracts\Document\LinkInterface;
+use Neomerx\JsonApi\Contracts\Encoder\Parameters\EncodingParametersInterface;
 use Neomerx\JsonApi\Contracts\Schema\ContainerInterface as SchemaContainerInterface;
 use Neomerx\JsonApi\Encoder\EncoderOptions;
-use Neomerx\JsonApi\Exceptions\ErrorCollection;
 use Neomerx\JsonApi\Factories\Factory as BaseFactory;
-use Psr\Http\Message\MessageInterface;
+use Psr\Http\Message\RequestInterface as PsrRequest;
 use Psr\Http\Message\ResponseInterface as PsrResponse;
-use Psr\Http\Message\ServerRequestInterface;
 use function CloudCreativity\JsonApi\http_contains_body;
 use function CloudCreativity\JsonApi\json_decode;
 
@@ -80,25 +77,35 @@ class Factory extends BaseFactory implements FactoryInterface
     }
 
     /**
-     * @inheritDoc
+     * @inheritdoc
      */
-    public function createRequest(
-        ServerRequestInterface $httpRequest,
-        RequestInterpreterInterface $intepreter,
-        StoreInterface $store
+    public function createInboundRequest(
+        $method,
+        $resourceType,
+        $resourceId = null,
+        $relationshipName = null,
+        $relationships = false,
+        DocumentInterface $document = null,
+        EncodingParametersInterface $parameters = null
     ) {
-        $requestFactory = new RequestFactory($this);
-
-        return $requestFactory->build($httpRequest, $intepreter, $store);
+        return new InboundRequest(
+            $method,
+            $resourceType,
+            $resourceId,
+            $relationshipName,
+            $relationships,
+            $document,
+            $parameters
+        );
     }
 
 
     /**
      * @inheritDoc
      */
-    public function createResponse(PsrResponse $response)
+    public function createResponse(PsrRequest $request, PsrResponse $response)
     {
-        return new Response($response, $this->createDocumentObject($response));
+        return new Response($response, $this->createDocumentObject($request, $response));
     }
 
     /**
@@ -112,13 +119,13 @@ class Factory extends BaseFactory implements FactoryInterface
     /**
      * @inheritDoc
      */
-    public function createDocumentObject(MessageInterface $message)
+    public function createDocumentObject(PsrRequest $request, PsrResponse $response = null)
     {
-        if (!http_contains_body($message)) {
+        if (!http_contains_body($request, $response)) {
             return null;
         }
 
-        return new Document(json_decode($message->getBody()));
+        return new Document(json_decode($response ? $response->getBody() : $request->getBody()));
     }
 
     /**
@@ -146,20 +153,9 @@ class Factory extends BaseFactory implements FactoryInterface
     /**
      * @inheritdoc
      */
-    public function createStore(AdapterContainerInterface $adapters)
+    public function createStore(ContainerInterface $container)
     {
-        return new Store($adapters);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function createAdapterContainer(array $adapters)
-    {
-        $container = new AdapterContainer();
-        $container->registerMany($adapters);
-
-        return $container;
+        return new Store($container);
     }
 
     /**