+++ src/Store/Store.php
@@ -18,12 +18,17 @@
 
 namespace CloudCreativity\JsonApi\Store;
 
+use CloudCreativity\JsonApi\Contracts\Adapter\HasManyAdapterInterface;
+use CloudCreativity\JsonApi\Contracts\ContainerInterface;
+use CloudCreativity\JsonApi\Contracts\Object\RelationshipInterface;
+use CloudCreativity\JsonApi\Contracts\Object\ResourceIdentifierCollectionInterface;
 use CloudCreativity\JsonApi\Contracts\Object\ResourceIdentifierInterface;
-use CloudCreativity\JsonApi\Contracts\Store\AdapterInterface;
-use CloudCreativity\JsonApi\Contracts\Store\ContainerInterface;
+use CloudCreativity\JsonApi\Contracts\Object\ResourceObjectInterface;
+use CloudCreativity\JsonApi\Contracts\Store\StoreAwareInterface;
 use CloudCreativity\JsonApi\Contracts\Store\StoreInterface;
 use CloudCreativity\JsonApi\Exceptions\RecordNotFoundException;
 use CloudCreativity\JsonApi\Exceptions\RuntimeException;
+use CloudCreativity\JsonApi\Object\ResourceIdentifier;
 use Neomerx\JsonApi\Contracts\Encoder\Parameters\EncodingParametersInterface;
 
 /**
@@ -66,11 +71,135 @@ class Store implements StoreInterface
     /**
      * @inheritDoc
      */
-    public function query($resourceType, EncodingParametersInterface $params)
+    public function queryRecords($resourceType, EncodingParametersInterface $params)
     {
         return $this
             ->adapterFor($resourceType)
             ->query($params);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function createRecord($resourceType, ResourceObjectInterface $resource, EncodingParametersInterface $params)
+    {
+        $record = $this
+            ->adapterFor($resourceType)
+            ->create($resource, $params);
+
+        if ($schema = $this->container->getSchemaByResourceType($resourceType)) {
+            $identifier = ResourceIdentifier::create($resourceType, $schema->getId($record));
+            $this->identityMap->add($identifier, $record);
+        }
+
+        return $record;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function readRecord($resourceType, $resourceId, EncodingParametersInterface $params)
+    {
+        $record = $this
+            ->adapterFor($resourceType)
+            ->read($resourceId, $params);
+
+        $this->identityMap->add(ResourceIdentifier::create($resourceType, $resourceId), $record ?: false);
+
+        return $record;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function updateRecord($record, ResourceObjectInterface $resource, EncodingParametersInterface $params)
+    {
+        return $this
+            ->adapterFor($record)
+            ->update($record, $resource, $params);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function deleteRecord($record, EncodingParametersInterface $params)
+    {
+        $adapter = $this->adapterFor($record);
+
+        if (!$adapter->delete($record, $params)) {
+            throw new RuntimeException('Record could not be deleted.');
+        }
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function queryRelated(
+        $record,
+        $relationshipName,
+        EncodingParametersInterface $params
+    ) {
+        return $this
+            ->adapterFor($record)
+            ->related($relationshipName)
+            ->query($record, $params);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function queryRelationship(
+        $record,
+        $relationshipName,
+        EncodingParametersInterface $params
+    ) {
+        return $this
+            ->adapterFor($record)
+            ->related($relationshipName)
+            ->relationship($record, $params);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function replaceRelationship(
+        $record,
+        $relationshipName,
+        RelationshipInterface $relationship,
+        EncodingParametersInterface $params
+    ) {
+        return $this
+            ->adapterFor($record)
+            ->related($relationshipName)
+            ->replace($record, $relationship, $params);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function addToRelationship(
+        $record,
+        $relationshipName,
+        RelationshipInterface $relationship,
+        EncodingParametersInterface $params
+    ) {
+        return $this
+            ->adapterForHasMany($record, $relationshipName)
+            ->add($record, $relationship, $params);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function removeFromRelationship(
+        $record,
+        $relationshipName,
+        RelationshipInterface $relationship,
+        EncodingParametersInterface $params
+    ) {
+        return $this
+            ->adapterForHasMany($record, $relationshipName)
+            ->remove($record, $relationship, $params);
     }
 
     /**
@@ -118,7 +247,7 @@ class Store implements StoreInterface
     /**
      * @inheritdoc
      */
-    public function findRecord(ResourceIdentifierInterface $identifier)
+    public function findOrFail(ResourceIdentifierInterface $identifier)
     {
         if (!$record = $this->find($identifier)) {
             throw new RecordNotFoundException($identifier);
@@ -128,15 +257,62 @@ class Store implements StoreInterface
     }
 
     /**
-     * @param $resourceType
-     * @return AdapterInterface
+     * @inheritDoc
+     * @deprecated
      */
-    protected function adapterFor($resourceType)
+    public function findRecord(ResourceIdentifierInterface $identifier)
     {
+        return $this->findOrFail($identifier);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function findMany(ResourceIdentifierCollectionInterface $identifiers)
+    {
+        $results = [];
+
+        foreach ($identifiers->map() as $resourceType => $ids) {
+            $results = array_merge($results, $this->adapterFor($resourceType)->findMany($ids));
+        }
+
+        return $results;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function adapterFor($resourceType)
+    {
+        if (is_object($resourceType)) {
+            return $this->container->getAdapter($resourceType);
+        }
+
         if (!$adapter = $this->container->getAdapterByResourceType($resourceType)) {
             throw new RuntimeException("No adapter for resource type: $resourceType");
         }
 
+        if ($adapter instanceof StoreAwareInterface) {
+            $adapter->withStore($this);
+        }
+
         return $adapter;
     }
+
+    /**
+     * @param $resourceType
+     * @param $relationshipName
+     * @return HasManyAdapterInterface
+     */
+    private function adapterForHasMany($resourceType, $relationshipName)
+    {
+        $adapter = $this->adapterFor($resourceType)->related($relationshipName);
+
+        if (!$adapter instanceof HasManyAdapterInterface) {
+            throw new RuntimeException("Expecting a has-many relationship adapter.");
+        }
+
+        return $adapter;
+    }
+
 }